refactor: move config CLI to function
