+++ src/cli/mod.rs
@@ -5,7 +5,10 @@ use clap::{Parser, Subcommand};
 use clap_verbosity_flag::{self, Verbosity};
 use config::Config;
 
+
 use get::GetArgs;
+
+
 
 #[derive(Parser, Debug)]
 #[command(author, version)]
